lab2/requirements.txt
fastapi==0.95.1
uvicorn==0.18.3
pydantic==1.10.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5